+++ transaction-service/src/main/java/com/github/jukkarol/dto/accountDto/response/CreateAccountResponse.java
@@ -1,7 +1,9 @@
 package com.github.jukkarol.dto.accountDto.response;
 
+import java.math.BigDecimal;
+
 public record CreateAccountResponse(
         String accountNumber,
 
-        Integer  balance
+        BigDecimal balance
 ) { }
