+++ transaction-service/src/main/java/com/github/jukkarol/dto/depositDto/event/DepositRequestEvent.java
@@ -1,7 +1,9 @@
 package com.github.jukkarol.dto.depositDto.event;
 
+import java.math.BigDecimal;
+
 public record DepositRequestEvent(
-        Integer amount,
+        BigDecimal amount,
 
         String  accountNumber
 ) { }
