+++ transaction-service/src/main/java/com/github/jukkarol/dto/transactionDto/request/MakeTransactionRequest.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record MakeTransactionRequest(
         @JsonIgnore
         Long userId,
@@ -19,5 +21,5 @@ public record MakeTransactionRequest(
 
         @NotNull
         @Positive
-        Integer amount
+        BigDecimal amount
 ) { }
